requirements.txt
aiohttp==3.*
coloredlogs
dask
distributed
dask-kubernetes
kubernetes
pyzmq
psycopg2
requests
sqlalchemy
tornado
transaction
pyjwt>2
pyyaml
voluptuous
databases
fastapi
uvicorn
aiosqlite
alembic
